Assistant UI Example fixes #2
